[refactor] 기존 컨트롤러/클라이언트 클래스 추가
